⭕(wip): refactor shipment form
#360
