+++ ui/src/app/shipment/_components/form/commodity/commodity-dialog.tsx
@@ -129,7 +129,7 @@ function CommodityForm({ index }: { index: number }) {
   const { control, setValue } = useFormContext<ShipmentSchema>();
 
   return (
-    <FormGroup>
+    <CommodityFormInner>
       <FormControl>
         <CommodityAutocompleteField<ShipmentSchema>
           name={`commodities.${index}.commodityId`}
@@ -169,6 +169,10 @@ function CommodityForm({ index }: { index: number }) {
           description="Enter the weight of a single piece of this commodity."
         />
       </FormControl>
-    </FormGroup>
+    </CommodityFormInner>
   );
+}
+
+function CommodityFormInner({ children }: { children: React.ReactNode }) {
+  return <FormGroup>{children}</FormGroup>;
 }
